+++ ui/package.json
@@ -29,6 +29,7 @@
     "start:dev": "react-scripts start",
     "start:serve": "serve -s build -l 3000",
     "build": "react-scripts build",
+    "build:vis": "npm run build && npm run start:serve",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
